Introduce manual composition root resolving starter
