+++ SimpleContainer.Unity/Assets/SimpleContainer.Unity/Roots/UnityProjectRootBase.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 38fe4363429b44f9ad4c7c531de0c1f2
+timeCreated: 1580365806
