+++ SimpleContainer.Unity/Assets/SimpleContainer.Unity/Roots/UnityProjectRootManual.cs
@@ -1,0 +1,10 @@
+﻿namespace SimpleContainer.Unity.Roots
+{
+    public sealed class UnityProjectRootManual : UnityProjectRootBase
+    {
+        public async void InstallAsync()
+        {
+            await InstallAsyncInternally();
+        }
+    }
+}
