+++ SimpleContainer.Unity/Assets/SimpleContainer.Unity/Roots/UnityProjectRootManual.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 02abd9122cff4a85948cf95ff0ea577a
+timeCreated: 1580365800
