+++ SimpleContainer.Unity/Assets/SimpleContainer.Unity/Roots/UnityProjectRootOnAwake.cs
@@ -1,0 +1,10 @@
+﻿namespace SimpleContainer.Unity.Roots
+{
+    public class UnityProjectRootOnAwake : UnityProjectRootBase
+    {
+        async void Awake()
+        {
+            await InstallAsyncInternally();
+        }
+    }
+}
