+++ SimpleContainer.Unity/Assets/SimpleContainer.Unity/Roots/UnityProjectRootOnAwake.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 70dee5dc2414444686fad698e1559b3a
+timeCreated: 1534193537
